Add files via upload
Voice Assistant is a Python-based application that interacts with users through speech recognition and synthesis. It allows users to perform tasks such as checking the time, opening Google Chrome, playing music from YouTube, and browsing the web. The program integrates speech-to-text conversion using speech_recognition, text-to-speech synthesis with pyttsx3, and music playback capabilities with pywhatkit and pygame. It offers a seamless user experience with clear audio feedback and handles various commands effectively, making it a versatile tool for everyday tasks.